handle multiple env in onboarings page
