--- src/pages/onboardingsPage/components/OnboardingsTable/tableColumns.tsx
@@ -1,360 +1,0 @@
-import { ArrowForward } from '@mui/icons-material';
-import SentimentDissatisfiedIcon from '@mui/icons-material/SentimentDissatisfied';
-import { Box, Chip, Stack, Tooltip, Typography } from '@mui/material';
-import { GridColDef, GridOverlay, GridRenderCellParams } from '@mui/x-data-grid';
-import { ButtonNaked } from '@pagopa/mui-italia';
-import { usePermissions } from '@pagopa/selfcare-common-frontend/lib';
-import { trackEvent } from '@pagopa/selfcare-common-frontend/lib/services/analyticsService';
-import { Actions } from '@pagopa/selfcare-common-frontend/lib/utils/constants';
-import { resolvePathVariables } from '@pagopa/selfcare-common-frontend/lib/utils/routes-utils';
-import { TFunction } from 'i18next';
-import { useState } from 'react';
-import { Trans, useTranslation } from 'react-i18next';
-import { useHistory, useLocation } from 'react-router-dom';
-import { OnboardingIndexResource } from '../../../../api/generated/party-registry-proxy/OnboardingIndexResource';
-import BackofficeNotIntegratedModal from '../../../../components/BackofficeNotIntegratedModal';
-import { useTokenExchange } from '../../../../hooks/useTokenExchange';
-import { Product } from '../../../../model/Product';
-import { STATUSES_ALLOWED_TO_SEE_REQUESTS } from '../../../../utils/constants';
-import { ENV } from '../../../../utils/env';
-import { isProductAllowed } from '../../../adminPage/utils/utils';
-
-const STATUS_CHIP_CONFIG: Record<
-  string,
-  {
-    label: string;
-    tooltipText: string;
-    color: 'success' | 'warning' | 'error' | 'info' | 'default';
-  }
-> = {
-  TOBEVALIDATED: {
-    label: 'Da validare',
-    tooltipText: 'Richiesta di adesione in attesa di approvazione da parte di un utente PagoPA.',
-    color: 'default',
-  },
-  PENDING: {
-    label: 'In attesa',
-    tooltipText:
-      "L'ente deve caricare il contratto firmato per completare la richiesta di adesione.",
-    color: 'warning',
-  },
-  COMPLETED: {
-    label: 'Attivo',
-    tooltipText: "L'ente è aderente al prodotto. La richiesta è stata completata con successo.",
-    color: 'success',
-  },
-  REJECTED: {
-    label: 'Rifiutato',
-    tooltipText: 'Richiesta di adesione rifiutata da un utente PagoPA dopo le verifiche.',
-    color: 'error',
-  },
-  DELETED: {
-    label: 'Disattivo',
-    tooltipText: "L'ente non risulta più aderente al prodotto.",
-    color: 'error',
-  },
-  FAILED: {
-    label: 'In errore',
-    tooltipText: 'La richiesta di adesione non è andata a buon fine a causa di un errore tecnico.',
-    color: 'error',
-  },
-  EXPIRED: {
-    label: 'Scaduta',
-    tooltipText: 'La richiesta non è stata completata entro i termini',
-    color: 'error',
-  },
-};
-
-const truncatedCellSx = {
-  overflow: 'hidden',
-  textOverflow: 'ellipsis',
-  whiteSpace: 'nowrap',
-  width: '100%',
-  pl: 1,
-} as const;
-
-const renderCellWithTooltip = (params: GridRenderCellParams) => {
-  const text = params.formattedValue ?? params.value ?? '';
-  return (
-    <Tooltip title={text} arrow enterDelay={300}>
-      <Box sx={truncatedCellSx}>{text}</Box>
-    </Tooltip>
-  );
-};
-
-const renderUpdatedAtCell = (params: GridRenderCellParams<Date>) => {
-  const date = params.value;
-  const text = date ? date.toLocaleDateString() : '-';
-  return (
-    <Tooltip title={text} arrow enterDelay={300}>
-      <Box sx={truncatedCellSx}>{text}</Box>
-    </Tooltip>
-  );
-};
-
-const renderStatusCell = (params: GridRenderCellParams<string>) => {
-  const value = params.value ?? '';
-  const config = STATUS_CHIP_CONFIG[value] ?? {
-    label: value,
-    tooltipText: '',
-    color: 'default' as const,
-  };
-  return (
-    <Tooltip title={config.tooltipText} arrow enterDelay={300}>
-      <Chip label={config.label} color={config.color} size="small" sx={{ ml: 1 }} />
-    </Tooltip>
-  );
-};
-
-const renderHeaderWithTooltip = (params: any) => (
-  <Tooltip title={params.colDef.description || params.colDef.headerName} arrow enterDelay={300}>
-    <Box sx={{ overflow: 'hidden', textOverflow: 'ellipsis', whiteSpace: 'nowrap', width: '100%' }}>
-      {params.colDef.headerName}
-    </Box>
-  </Tooltip>
-);
-
-const DescriptionCell = ({ params }: { params: GridRenderCellParams<OnboardingIndexResource> }) => {
-  const history = useHistory();
-  const text = params.row?.description ?? '';
-  const isClickable = params.row?.status === 'COMPLETED';
-
-  if (!isClickable) {
-    return (
-      <Tooltip title={text} arrow enterDelay={300}>
-        <Box sx={truncatedCellSx}>{text}</Box>
-      </Tooltip>
-    );
-  }
-
-  return (
-    <Tooltip title={text} arrow enterDelay={300}>
-      <Box sx={truncatedCellSx}>
-        <ButtonNaked
-          color="primary"
-          component="button"
-          sx={{
-            fontWeight: 700,
-            fontSize: '16px',
-            textAlign: 'left',
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-            whiteSpace: 'nowrap',
-            maxWidth: '100%',
-            display: 'block',
-          }}
-          onClick={() => {
-            history.push(
-              resolvePathVariables(ENV.ROUTES.ADMIN_ONBOARDINGS_DETAIL, {
-                partyId: params.row.institutionId,
-              })
-            );
-          }}
-        >
-          {text}
-        </ButtonNaked>
-      </Box>
-    </Tooltip>
-  );
-};
-
-const ActionCell = ({
-  params,
-  products,
-}: {
-  params: GridRenderCellParams<OnboardingIndexResource>;
-  products: Array<Product>;
-}) => {
-  const { t } = useTranslation();
-  const history = useHistory();
-  const { hasPermission } = usePermissions();
-  const { invokeProductBo } = useTokenExchange();
-  const [openModal, setOpenModal] = useState(false);
-
-  const productId = params.row?.productId || '';
-  const status = params.row?.status;
-
-  const canAccessBackofficeAdmin =
-    hasPermission(productId, Actions.AccessProductBackofficeAdmin) ||
-    hasPermission('ALL', Actions.AccessProductBackofficeAdmin);
-
-  const canAccessAccountPage =
-    hasPermission(productId, Actions.ViewAccountPage) ||
-    (hasPermission('ALL', Actions.ViewAccountPage) &&
-      STATUSES_ALLOWED_TO_SEE_REQUESTS.includes(status));
-
-  const productName = products.find((p) => p.id === productId)?.title || productId;
-
-  return (
-    <>
-      {status === 'COMPLETED' && canAccessBackofficeAdmin && (
-        <ButtonNaked
-          component="button"
-          endIcon={<ArrowForward />}
-          onClick={() => {
-            trackEvent('BACKSTAGE_BACK_OFFICE_CLICK', {
-              product_id: productId,
-            });
-            if (isProductAllowed(productId)) {
-              const selectedParty = params.row;
-              const productFromConfiguration = products.find((p) => p.id === productId);
-              if (productFromConfiguration) {
-                void invokeProductBo(productFromConfiguration, selectedParty);
-              }
-            } else {
-              setOpenModal(true);
-            }
-          }}
-          sx={{ color: 'primary.main', fontWeight: 'bold', mr: 2 }}
-        >
-          {t('adminPage.selectedPartyDetails.backOffice')}
-        </ButtonNaked>
-      )}
-      {!canAccessBackofficeAdmin && canAccessAccountPage && (
-        <ButtonNaked
-          component="button"
-          endIcon={<ArrowForward />}
-          onClick={() => {
-            history.push(
-              resolvePathVariables(ENV.ROUTES.ADMIN_REQUEST_DETAIL, {
-                tokenId: params.row.onboardingId,
-              }),
-              { fromDashboard: true }
-            );
-          }}
-          sx={{ color: 'primary.main', fontWeight: 'bold', mr: 2 }}
-        />
-      )}
-      <BackofficeNotIntegratedModal
-        open={openModal}
-        productName={productName}
-        onClose={() => setOpenModal(false)}
-      />
-    </>
-  );
-};
-
-export const RenderNoRowsOverlay = () => {
-  const history = useHistory();
-  const location = useLocation();
-  const { t } = useTranslation();
-
-  const handleReset = () => {
-    history.push({
-      pathname: location.pathname,
-      search: '',
-    });
-  };
-
-  return (
-    <GridOverlay sx={{ pointerEvents: 'auto' }}>
-      <Stack
-        alignItems="center"
-        justifyContent="center"
-        flexDirection="row"
-        sx={{ height: '100%', width: '100%', pointerEvents: 'auto' }}
-      >
-        <SentimentDissatisfiedIcon />
-        <Typography variant="body2" sx={{ display: 'flex', alignItems: 'center', ml: 1 }}>
-          <Trans i18nKey="onboardingsPage.table.noResults">
-            I filtri che hai applicato non hanno dato nessun risultato.
-          </Trans>
-        </Typography>
-        <ButtonNaked
-          size="small"
-          onClick={handleReset}
-          sx={{ ml: 1, color: '#0B3EE3', cursor: 'pointer', pointerEvents: 'auto' }}
-        >
-          {t('onboardingsPage.filters.resetButton')}
-        </ButtonNaked>
-      </Stack>
-    </GridOverlay>
-  );
-};
-
-export const getOnboardingsColumns = (
-  t: TFunction,
-  products: Array<Product>,
-  hasBackofficeAdmin: boolean
-): Array<GridColDef<OnboardingIndexResource>> => [
-    {
-      field: 'description',
-      headerName: t('onboardingsPage.table.institutionName'),
-      flex: 2,
-      sortable: false,
-      valueGetter: (params) => {
-        const description = params.row?.description;
-        const parentDescription = params.row?.parentDescription;
-        return description && parentDescription
-          ? `${description} - ${parentDescription}`
-          : description || '';
-      },
-      renderCell: (params) => <DescriptionCell params={params} />,
-    },
-    {
-      field: 'productId',
-      headerName: t('onboardingsPage.table.product'),
-      flex: 1,
-      sortable: false,
-      valueGetter: (params) => {
-        const productId = params.row?.productId;
-
-        const subProduct = products
-          .flatMap((p) => p.subProducts ?? [])
-          .find((sp) => sp.id === productId);
-
-        const product = products.find((p) => p.id === productId);
-
-        return subProduct?.title ?? product?.title ?? productId ?? '';
-      },
-      renderCell: renderCellWithTooltip,
-    },
-    {
-      field: 'institutionType',
-      headerName: t('onboardingsPage.table.institutionType'),
-      description: t('onboardingsPage.table.institutionTypeTooltip'),
-      flex: 1.5,
-      sortable: false,
-      renderHeader: renderHeaderWithTooltip,
-      valueGetter: (params) => {
-        const raw = params.row?.institutionType;
-        if (!raw) {
-          return '';
-        }
-        const key = `common.institutionType.descriptions.${raw.toLowerCase()}`;
-        const translated = t(key);
-        return translated === key ? raw : translated;
-      },
-      renderCell: renderCellWithTooltip,
-    },
-    {
-      field: 'requestDate',
-      headerName: t('onboardingsPage.table.requestDate'),
-      description: t('onboardingsPage.table.requestDateTooltip'),
-      flex: 1,
-      sortable: true,
-      renderHeader: renderHeaderWithTooltip,
-      valueGetter: (params) => {
-        const date = params.row?.createdAt;
-        return date ? new Date(date) : null;
-      },
-      renderCell: renderUpdatedAtCell,
-    },
-    {
-      field: 'status',
-      headerName: t('onboardingsPage.table.status'),
-      flex: 1,
-      sortable: false,
-      renderCell: renderStatusCell,
-    },
-    {
-      field: 'actions',
-      headerName: '',
-      width: hasBackofficeAdmin ? 160 : 56,
-      sortable: false,
-      disableColumnMenu: true,
-      align: 'right',
-      headerAlign: 'right',
-      renderCell: (params) => <ActionCell params={params} products={products} />,
-    },
-  ];
